Flytter inline komponenter ut fra Barnekort inn i egne komponenter
